Task6 Commited

+++ 0x00-hello_world/6-size.c
@@ -11,10 +11,10 @@ int main(void)
 	long long int e;
 	float e;
 
-	printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(a));
-	printf("Size of an int: %lu byte(s)\n", (unsigned long)sizeof(b));
-	printf("Size of long int: %lu byte(s)\n", (unsigned long)sizeof(c));
-	printf("Size of long long int: %lu byte(s)\n", (unsigned long)sizeof(d));
-	printf("Size of float: %lu byte(s)\n", (unsigned long)sizeof(e));
+	printf("Size of a char: %zu byte(s)\n", sizeof(a));
+	printf("Size of an int: %zu byte(s)\n", sizeof(b));
+	printf("Size of long int: %zu byte(s)\n", sizeof(c));
+	printf("Size of long long int: %zu byte(s)\n", sizeof(d));
+	printf("Size of float: %zu byte(s)\n", sizeof(e));
 	return (0);
 }
